Fix elevation compute. Dont use readBitmapColor with floats as sometimes the average can give wrong rounded value and thus wrong elevation. This fixes -5000 elevations

+++ all/native/rastertiles/MapBoxElevationDataDecoder.cpp
@@ -12,7 +12,7 @@ namespace carto
     {
     }
 
-    std::array<float, 4> MapBoxElevationDataDecoder::getColorComponentCoefficients() const
+    std::array<double, 4> MapBoxElevationDataDecoder::getColorComponentCoefficients() const
     {
         return COMPONENTS;
     }
@@ -22,6 +22,6 @@ namespace carto
         return SCALES;
     }
 
-    const std::array<float, 4> MapBoxElevationDataDecoder::COMPONENTS = std::array<float, 4>{256 * 256 * 0.1f, 256 * 0.1f, 0.1f, -10000.0f};
+    const std::array<double, 4> MapBoxElevationDataDecoder::COMPONENTS = std::array<double, 4>{256 * 256 * 0.1f, 256 * 0.1f, 0.1f, -10000.0f};
     const std::array<float, 4> MapBoxElevationDataDecoder::SCALES = std::array<float, 4>{256 * 256, 256.0f, 1.0f, 0.0f};
 } // namespace carto
